+++ onnx2torch/node_converters/binary_math_operations.py
@@ -67,53 +67,3 @@ def _(node: OnnxNode, graph: OnnxGraph) -> OperationConverterResult:  # pylint: 
         ),
         onnx_mapping=onnx_mapping_from_node(node=node),
     )
-
-
-class OnnxNeg(nn.Module, OnnxToTorchModule):
-    def forward(self, x):
-        return -x
-
-    
-@add_converter(operation_type='Neg', version=1)
-@add_converter(operation_type='Neg', version=6)
-@add_converter(operation_type='Neg', version=13)
-def _(node: OnnxNode, graph: OnnxGraph) -> OperationConverterResult:  # pylint: disable=unused-argument
-    return OperationConverterResult(
-        torch_module=OnnxNeg(),
-        onnx_mapping=onnx_mapping_from_node(node=node),
-    )
-
-
-class OnnxReciprocal(nn.Module, OnnxToTorchModule):
-    def forward(self, x):
-        return torch.reciprocal(x)
-
-    
-@add_converter(operation_type='Reciprocal', version=1)
-@add_converter(operation_type='Reciprocal', version=6)
-@add_converter(operation_type='Reciprocal', version=13)
-def _(node: OnnxNode, graph: OnnxGraph) -> OperationConverterResult:  # pylint: disable=unused-argument
-    return OperationConverterResult(
-        torch_module=OnnxReciprocal(),
-        onnx_mapping=onnx_mapping_from_node(node=node),
-    )
-
-
-class OnnxEinsum(nn.Module, OnnxToTorchModule):
-
-    def __init__(self, equation: str):
-        super().__init__()
-        self.equation = equation
-        
-    def forward(self, *args):
-        return torch.einsum(self.equation, *args)
-
-
-@add_converter(operation_type='Einsum', version=12)
-def _(node: OnnxNode, graph: OnnxGraph) -> OperationConverterResult:  # pylint: disable=unused-argument
-    return OperationConverterResult(
-        torch_module=OnnxEinsum(
-            equation=node.attributes['equation'],
-        ),
-        onnx_mapping=onnx_mapping_from_node(node=node),
-    )
